Change NavBar notification to show only software updates
